Move Speed_Bumps class into main plugin file
Adds the autoloader to the Speed_Bumps class, and organizes both
together inside the main plugin file for easier reading.
